app/requirements.txt
requests
python-dotenv
sendgrid==6.05